Prestige Lavender Fields Whitefield Bangalore
Prestige Lavender Fields Whitefield is a newly launched residential project that will let you settle in the greenery that gives you the gift of modern living. The spacious layouts are the best and are made keeping in mind the optimized space utilization. All the flats are an absolute pleasure to live in and especially all the corner flats are well-ventilated and have natural light facilities.

Whitefield Bangalore is located in close proximity further enhancing the connectivity. Apart from this, NH-207 is another well-laid-out network that gives a boost to the lives of the residents.

Now, you can gather and enjoy family time in the spacious living rooms, feel comfortable and relax your senses in the exclusive bedrooms. Enhance your lifestyle with warm and luxury amenities tailored just for you. The family that eats together, stays together and thus a bigger dining area will lead to more love. Make cooking fun with impressive built-in appliances, lighting, and an open-format kitchen.

Prestige Lavender Fields offers the best floor plans for 1 BHK, 2 BHK, and 3 BHK luxury residential accommodations.  This dream project will not only match your standard of living but also promote enhanced living that is in sync with your status. Impressive design, round-the-clock safety and security, wide open green spaces, and spectacular views of the surroundings are some that make this the top residential project in Whitefield Bangalore.

It sets new benchmarks for other developers in the industry in terms of luxurious and world-class amenities. It is enriched with badminton courts, a fitness center; A temperature-controlled indoor pool, a jogging track, an outdoor play area, and a clubhouse. Attaining maximum customer satisfaction with its unique architecture Prestige Lavender Fields Whitefield Bengaluru welcomes lavish living.

For More Details
Visit us - https://bit.ly/3uv5pUk